jittor_rt_detr/requirements.txt
jittor>=1.3.0
numpy>=1.21.0
scipy>=1.7.0
Pillow>=8.0.0
tqdm>=4.60.0
matplotlib>=3.5.0
opencv-python>=4.5.0
pycocotools>=2.0.0 